requirement.txt
flask
gunicorn
sklearn
pandas
numpy
matplotlib
PyYAML
evidently
dill
imblearn
patool
pyunpack
-e .